atomic internal tx

+++ src/services/main/paymentManager.ts
@@ -6,7 +6,7 @@ import { InboundOptionals, defaultInvoiceExpiry } from '../storage/paymentStorag
 import LND from '../lnd/lnd.js'
 import { Application } from '../storage/entity/Application.js'
 import { ERROR, getLogger, PubLogger } from '../helpers/logger.js'
-import { AddressPaidCb, InvoicePaidCb, NewBlockCb } from '../lnd/settings.js'
+import { AddressPaidCb, InvoicePaidCb, NewBlockCb, TxOutput } from '../lnd/settings.js'
 import { UserReceivingInvoice, ZapInfo } from '../storage/entity/UserReceivingInvoice.js'
 import { Payment_PaymentStatus } from '../../../proto/lnd/lightning.js'
 import { Event, verifiedSymbol, verifyEvent } from 'nostr-tools'
@@ -578,8 +578,8 @@ export default class {
     }
 
     async PayInternalInvoice(userId: string, internalInvoice: UserReceivingInvoice, amounts: { payAmount: number, serviceFee: number }, linkedApplication: Application, debitNpub?: string) {
-        if (amounts.payAmount < 0) {
-            throw new Error("amount cannot be negative")
+        if (amounts.payAmount <= 0) {
+            throw new Error("amount cannot be zero or negative")
         }
         if (internalInvoice.paid_at_unix > 0) {
             throw new Error("this invoice was already paid")
@@ -612,8 +612,8 @@ export default class {
 
 
     async CreditIncomingInvoice(invoice: string, amount: number, internal: boolean, txId: string) {
-        if (amount < 0) {
-            throw new Error("amount cannot be negative")
+        if (amount <= 0) {
+            throw new Error("amount cannot be zero or negative")
         }
         const userInvoice = await this.storage.paymentStorage.GetInvoiceOwner(invoice, txId)
         if (!userInvoice) {
@@ -688,8 +688,13 @@ export default class {
         let amount = req.amountSats
         if (req.swap_operation_id) {
             const swap = await this.storage.paymentStorage.GetTransactionSwap(req.swap_operation_id, ctx.app_user_id)
-            amount = amount > 0 ? amount : swap?.invoice_amount || 0
+            if (!swap) {
+                throw new Error("swap quote not found")
+            }
             await this.storage.paymentStorage.DeleteTransactionSwap(req.swap_operation_id)
+            if (amount <= 0) {
+                amount = swap.transaction_amount - swap.chain_fee_sats
+            }
         }
         if (amount <= 0) {
             throw new Error("invalid tx amount")
@@ -697,27 +702,61 @@ export default class {
         const { blockHeight } = await this.lnd.GetInfo()
         const app = await this.storage.applicationStorage.GetApplication(ctx.app_id)
         const isManagedUser = ctx.user_id !== app.owner.user_id
-        const serviceFee = this.getSendServiceFee(Types.UserOperationType.OUTGOING_USER_TO_USER, req.amountSats, isManagedUser)
+        const serviceFee = this.getSendServiceFee(Types.UserOperationType.OUTGOING_USER_TO_USER, amount, isManagedUser)
 
         const txId = crypto.randomBytes(32).toString("hex")
         const addressData = `${req.address}:${txId}`
-        await this.storage.userStorage.DecrementUserBalance(ctx.user_id, req.amountSats + serviceFee, addressData)
-        this.addressPaidCb({ hash: txId, index: 0 }, req.address, req.amountSats, 'internal')
-        if (isManagedUser && serviceFee > 0) {
-            await this.storage.userStorage.IncrementUserBalance(app.owner.user_id, serviceFee, 'fees')
+        const { newTx, receivingTx } = await this.storage.StartTransaction(async tx => {
+            await this.storage.userStorage.DecrementUserBalance(ctx.user_id, amount + serviceFee, addressData, tx)
+            const receivingTx = await this.CreditAddress(req.address, { hash: txId, index: 0 }, amount, true, blockHeight, tx)
+            if (isManagedUser && serviceFee > 0) {
+                await this.storage.userStorage.IncrementUserBalance(app.owner.user_id, serviceFee, 'fees', tx)
+            }
+            const chainFees = 0
+            const internalAddress = true
+            const newTx = await this.storage.paymentStorage.AddUserTransactionPayment(ctx.user_id, req.address, txId, 0, amount, chainFees, serviceFee, internalAddress, blockHeight, app, tx)
+            return { newTx, receivingTx }
+        })
+        const operationId = `${Types.UserOperationType.INCOMING_TX}-${receivingTx.serial_id}`
+        const op = { amount, paidAtUnix: Date.now() / 1000, inbound: true, type: Types.UserOperationType.INCOMING_TX, identifier: req.address, operationId, network_fee: 0, service_fee: receivingTx.service_fee, confirmed: true, tx_hash: txId, internal: true }
+        try {
+            await this.paymentSideEffects.sendOperationToNostr(receivingTx.user_address.linkedApplication!, receivingTx.user_address.user.user_id, op)
+        } catch (err: any) {
+            this.log(ERROR, "error sending operation to nostr for incoming tx", err.message || "")
         }
-        const chainFees = 0
-        const internalAddress = true
-        const newTx = await this.storage.paymentStorage.AddUserTransactionPayment(ctx.user_id, req.address, txId, 0, req.amountSats, chainFees, serviceFee, internalAddress, blockHeight, app)
+        this.utils.stateBundler.AddTxPoint('addressWasPaid', amount, { used: 'internal', from: 'system', timeDiscount: true }, receivingTx.user_address.linkedApplication!.app_id)
         const user = await this.storage.userStorage.GetUser(ctx.user_id)
         const txData = `${newTx.address}:${newTx.tx_hash}`
-        this.storage.eventsLog.LogEvent({ type: 'address_payment', userId: ctx.user_id, appId: app.app_id, appUserId: "", balance: user.balance_sats, data: txData, amount: req.amountSats })
+        this.storage.eventsLog.LogEvent({ type: 'address_payment', userId: ctx.user_id, appId: app.app_id, appUserId: "", balance: user.balance_sats, data: txData, amount })
         return {
             txId: txId,
             operation_id: `${Types.UserOperationType.OUTGOING_TX}-${newTx.serial_id}`,
-            network_fee: chainFees,
+            network_fee: 0,
             service_fee: serviceFee
         }
+    }
+
+    async CreditAddress(address: string, txOutput: TxOutput, amount: number, internal: boolean, txBroadcastHeight: number, dbTxId: string) {
+        const userAddress = await this.storage.paymentStorage.GetAddressOwner(address, dbTxId)
+        if (!userAddress) {
+            throw new Error("address not found")
+        }
+        if (!userAddress.linkedApplication) {
+            throw new Error("an address was paid, that has no linked application")
+        }
+
+        const isManagedUser = userAddress.user.user_id !== userAddress.linkedApplication.owner.user_id
+        const fee = this.getReceiveServiceFee(Types.UserOperationType.INCOMING_TX, amount, isManagedUser)
+        const txRecord = await this.storage.paymentStorage.AddAddressReceivingTransaction(userAddress, txOutput.hash, txOutput.index, amount, fee, internal, txBroadcastHeight, dbTxId)
+        if (internal) {
+            const addressData = `${userAddress.address}:${txOutput.hash}`
+            this.storage.eventsLog.LogEvent({ type: 'address_paid', userId: userAddress.user.user_id, appId: userAddress.linkedApplication!.app_id, appUserId: "", balance: userAddress.user.balance_sats, data: addressData, amount })
+            await this.storage.userStorage.IncrementUserBalance(userAddress.user.user_id, txRecord.paid_amount - fee, addressData, dbTxId)
+            if (fee > 0) {
+                await this.storage.userStorage.IncrementUserBalance(userAddress.linkedApplication!.owner.user_id, fee, 'fees', dbTxId)
+            }
+        }
+        return txRecord
     }
 
     async ListTxSwaps(ctx: Types.UserContext): Promise<Types.TxSwapsList> {
